+++ crates/objc2/src/macros/declare_class.rs
@@ -410,10 +410,6 @@ macro_rules! declare_class {
                 fn as_super(&self) {
                     &*self.__superclass
                 }
-
-                fn as_super_mut(&mut self) {
-                    &mut *self.__superclass
-                }
             }
         }
 
@@ -466,11 +462,6 @@ macro_rules! declare_class {
                 #[inline]
                 fn as_super(&self) -> &Self::Super {
                     &*self.__superclass
-                }
-
-                #[inline]
-                fn as_super_mut(&mut self) -> &mut Self::Super {
-                    &mut *self.__superclass
                 }
             }